requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.17,<3.18
drf-spectacular>=0.28,<0.29
django-cors-headers>=4.6,<4.8
openpyxl>=3.1,<3.2
psycopg[binary]>=3.3,<3.4
celery[redis]>=5.4,<5.6
redis>=5.0,<6.0
python-dateutil>=2.9,<3.0
requests>=2.32,<3.0
aiogram>=3.0,<4.0
aiohttp>=3.9,<4.0
aiohttp-socks>=0.11.0